% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrgdal-package.R
\docType{package}
\name{xrgdal-package}
\alias{xrgdal}
\alias{xrgdal-package}
\title{xrgdal: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mdsumner/xrgdal}
  \item Report bugs at \url{https://github.com/mdsumner/xrgdal/issues}
}

}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
